+++ app/Plugin/Admin/SettingsValidation/ScheduleInterval.php
@@ -10,6 +10,8 @@ namespace PersonioIntegrationLight\Plugin\Admin\SettingsValidation;
 // prevent direct access.
 defined( 'ABSPATH' ) || exit;
 
+use PersonioIntegrationLight\Plugin\Intervals;
+
 /**
  * Object which validates the given schedule interval.
  */
@@ -28,14 +30,19 @@ class ScheduleInterval {
 		// bail if the value is empty.
 		if ( empty( $value ) ) {
 			add_settings_error( $option, $option, __( 'An interval has to be set.', 'personio-integration-light' ) );
-			return '';
+
+			// return the old value.
+			return (string) get_option( $option );
 		}
 
 		// check if the given interval exists.
-		$intervals = wp_get_schedules();
+		$intervals = Intervals::get_instance()->get_intervals_for_settings();
 		if ( empty( $intervals[ $value ] ) ) {
 			/* translators: %1$s will be replaced by the name of the used interval */
 			add_settings_error( $option, $option, sprintf( __( 'The given interval %1$s does not exists.', 'personio-integration-light' ), esc_html( $value ) ) );
+
+			// return the old value.
+			return (string) get_option( $option );
 		}
 
 		// return the value.